refactor: move types into solving
