+++ expertsystem/amplitude/canonical_decay.py
@@ -4,7 +4,7 @@ from typing import Any, Callable, List, Optional
 
 from expertsystem.data import Spin
 from expertsystem.solving.topology import StateTransitionGraph
-from expertsystem.types import InteractionProperties, ParticleWithSpin
+from expertsystem.solving.types import InteractionProperties, ParticleWithSpin
 
 from .helicity_decay import (
     HelicityAmplitudeGenerator,
