+++ expertsystem/solving/__init__.py
@@ -29,7 +29,7 @@ from expertsystem.data import Parity, Particle, ParticleCollection, Spin
 from expertsystem.solving.conservation_rules import IsoSpinValidity, Rule
 from expertsystem.solving.properties import get_particle_property
 from expertsystem.solving.topology import StateTransitionGraph, Topology
-from expertsystem.types import (
+from expertsystem.solving.types import (
     EdgeQuantumNumber,
     EdgeQuantumNumbers,
     NodeQuantumNumber,
