+++ expertsystem/solving/conservation_rules.py
@@ -21,7 +21,7 @@ import attr
 
 from expertsystem.data import Spin
 from expertsystem.solving.combinatorics import arange
-from expertsystem.types import EdgeQuantumNumbers, NodeQuantumNumbers
+from expertsystem.solving.types import EdgeQuantumNumbers, NodeQuantumNumbers
 
 
 def is_boson(spin_magnitude: float) -> bool:
