+++ expertsystem/solving/types.py
