+++ expertsystem/ui/_default_settings.py
@@ -28,7 +28,7 @@ from expertsystem.solving.conservation_rules import (
     StrangenessConservation,
     TauLNConservation,
 )
-from expertsystem.types import EdgeQuantumNumbers, NodeQuantumNumbers
+from expertsystem.solving.types import EdgeQuantumNumbers, NodeQuantumNumbers
 
 EXPERT_SYSTEM_PATH = dirname(dirname(realpath(__file__)))
 DEFAULT_PARTICLE_LIST_FILE = "additional_particle_definitions.yml"
