+++ tests/unit/solving/conservation_rules/test_ls_coupling.py
@@ -2,7 +2,7 @@ import pytest
 
 from expertsystem.data import Spin
 from expertsystem.solving.conservation_rules import ParityConservationHelicity
-from expertsystem.types import InteractionProperties
+from expertsystem.solving.types import InteractionProperties
 from expertsystem.ui import InteractionTypes, StateTransitionManager
 from expertsystem.ui._default_settings import (
     create_default_interaction_settings,
